backend/web_agent.py
"""
Web Agent for Omni-Agent - Browser automation and web scraping functionality
"""
import asyncio
import json
from playwright.async_api import async_playwright


class WebAgent:
    def __init__(self):
        self.playwright = None
        self.browser = None
        self.page = None

    async def __aenter__(self):
        await self.start()
        return self

    async def __aexit__(self, exc_type, exc_val, exc_tb):
        await self.close()

    async def start(self):
        """Start the Playwright instance and browser"""
        self.playwright = await async_playwright().start()
        self.browser = await self.playwright.chromium.launch(headless=True)
        self.page = await self.browser.new_page()

    async def close(self):
        """Close the browser and Playwright instance"""
        if self.browser:
            await self.browser.close()
        if self.playwright:
            await self.playwright.stop()

    async def run_task(self, prompt, update_callback=None):
        """
        Run a web task based on the prompt
        
        Args:
            prompt: Natural language description of the task to perform
            update_callback: Optional callback to receive updates during execution
        
        Returns:
            Result of the web task
        """
        if update_callback:
            await update_callback(None, f"Starting web task: {prompt}")
        
        try:
            # Simple implementation - in a real scenario, you would parse the prompt
            # and perform appropriate web actions
            if "найди" in prompt.lower() or "search" in prompt.lower() or "find" in prompt.lower():
                # For demonstration, we'll just return a mock result
                # In a real implementation, you would actually navigate and search
                if update_callback:
                    await update_callback(None, "Navigating to search engine...")
                
                # Mock result - in real implementation, this would be actual search results
                result = f"Mock search results for: {prompt}"
                
                if update_callback:
                    await update_callback(None, f"Search completed. Found results for: {prompt}")
                
                return result
            else:
                # For other types of tasks, return a generic response
                return f"Processed web task: {prompt}. This is a simulated response from the web agent."
                
        except Exception as e:
            if update_callback:
                await update_callback(None, f"Error during web task: {str(e)}")
            return f"Error during web task: {str(e)}"


# Example usage
async def main():
    async with WebAgent() as agent:
        result = await agent.run_task("Найди погоду в Москве", 
                                      update_callback=lambda img, msg: print(f"Update: {msg}"))
        print(result)


if __name__ == "__main__":
    asyncio.run(main())